Split routing tests into one file per version
